+++ WP.API/Controllers/AuthController.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using WP.Services;
-using static WP.DTOs.UserDtos;
+using  WP.DTOs;
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
