+++ WP.Core/ITokenService.cs
@@ -3,7 +3,8 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using static WP.DTOs.UserDtos;
+using WP.DTOs;
+
 
 namespace WP.Core
 {
