+++ WP.Core/TokenService.cs
@@ -7,7 +7,8 @@ using System.Linq;
 using System.Security.Claims;
 using System.Text;
 using System.Threading.Tasks;
-using static WP.DTOs.UserDtos;
+using WP.DTOs;
+
 
 namespace WP.Core
 {
