+++ WP.Data/Repositories/IUserRepository.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using static WP.DTOs.UserDtos;
+using  WP.DTOs;
 
 namespace WP.Data.Repositories
 {
