+++ WP.Data/Repositories/UserRepository.cs
@@ -5,7 +5,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using static WP.DTOs.UserDtos;
+using  WP.DTOs;
 
 namespace WP.Data.Repositories
 {
